requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
alpaca-py==0.20.3
alpaca-trade-api==3.2.0
alpha-vantage==2.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
APScheduler==3.10.4
asttokens==2.4.1
attrs==23.2.0
Babel==2.14.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
cachetools==5.3.3
ccxt==4.2.85
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==41.0.5
cycler==0.12.1
dataclasses-json==0.6.5
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
duckdb==0.10.2
email_validator==2.1.1
exchange_calendars==4.5.3
executing==2.0.1
Flask==3.0.3
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==1.2.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.51.0
frozendict==2.4.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
ibapi==9.81.1.post1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
ipython==8.23.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpickle==3.0.4
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.50
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langsmith==0.1.53
lumibot==3.3.7
lumiwealth-tradier==0.1.7
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.3
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.25.1
orjson==3.10.2
packaging==23.2
pandas==2.2.2
pandas-datareader==0.10.0
pandas_market_calendars==4.4.0
paramiko==3.3.1
parso==0.8.4
passlib==1.7.4
peewee==3.17.3
pexpect==4.9.0
pillow==10.3.0
plotly==5.21.0
pluggy==1.5.0
polygon-api-client==1.13.5
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pycares==4.4.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.17.2
pyluach==2.2.0
PyNaCl==1.5.0
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
Quandl==3.7.0
quantstats-lumi==0.2.0
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scipy==1.13.0
seaborn==0.13.2
setuptools==69.5.1
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
speaklater==1.3
SQLAlchemy==2.0.29
sseclient-py==1.8.0
stack-data==0.6.3
streamlit==1.34.0
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.6.0
timedelta==2020.12.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
uuid==1.30
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.2
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
yfinance==0.2.38
wikipedia
numexpr
langchainhub
chromadb
docx2txt
pypdf
unstructured
python-pptx
nltk
boto3
llama-index-embeddings-bedrock
opensearch-py
requests_aws4auth
config
